requirements.txt
# Core app framework
streamlit==1.40.1

# Data handling
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0

# ML/NLP basics
scikit-learn==1.3.2
joblib==1.3.2
textblob==0.18.0
tqdm==4.66.5

# Visualization
plotly==5.19.0

# RAG & embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.15

# LLM API
openai==1.0.0

# File handling & parsing
PyPDF2==3.0.0
beautifulsoup4==4.12.0
